fixed exception handling bug when uploading invalid filetypes.
The a InvalidFileType Exception was raised (custom exception) which was
not caught in the outside functions. This causes an error 500 on the deployed version.

This error appeared in uploading measurements, pvmodules and individual
measurements.

Furthermore a legacy function in the measurement.py was removed.

The superfluous pass of the app.logger to a var applogger was also removed